pyproject.toml
[tool.poetry]
name = "test_backend"
version = "0.1.0"
description = ""
authors = ["qzibet <qzibet@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.5"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
